requirements.txt
asgiref==3.7.2
asn1crypto==1.5.1
attrs==23.1.0
black==23.12.1
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dj-database-url==2.1.0
Django==4.0.4
django-cuser==2017.3.16
django-debug-toolbar==3.4.0
django-filter==23.2
django-rest-framework==0.1.0
django-tinymce==3.6.1
djangorestframework==3.13.1
djangorestframework-jwt==1.0.1
djangorestframework-simplejwt==5.2.0
docker==7.0.0
drf-spectacular==0.22.1
flake8==5.0.4
flake8-quotes==3.3.1
flake8-variables-names==0.0.5
idna==3.6
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
Markdown==3.5.2
mccabe==0.7.0
modeltranslation==0.25
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
pep8-naming==0.13.2
pg8000==1.30.3
Pillow==10.1.0
platformdirs==4.1.0
pluggy==1.3.0
psycopg-binary==3.1.16
psycopg2-binary==2.9.9
pycodestyle==2.9.1
pyflakes==2.5.0
PyJWT==2.8.0
pytest==7.4.3
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0.1
referencing==0.32.0
requests==2.31.0
rpds-py==0.15.2
scramp==1.4.4
setuptools==69.0.3
six==1.16.0
sqlparse==0.4.4
testing.common.database==2.0.3
testing.postgresql==1.3.0
typing_extensions==4.9.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.2.1
wheel==0.42.0
